% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manage_addColumn.R
\name{manage_addColumn}
\alias{manage_addColumn}
\alias{mod_manage_addColumn_ui}
\alias{mod_manage_addColumn_server}
\title{manage_addColumn}
\usage{
mod_manage_addColumn_ui(id)

mod_manage_addColumn_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to add new columns to the loaded dataframe. It is a submodule of the manage module.

In particular, the user can programmatically add columns by writing down a custom \code{if_else} 
expression or by adding predefined calendar variables (\code{minute}, \code{hour}, \code{day}, \code{month}, 
\code{day_type}, \code{month_day}, \code{year}, \code{date}, \code{datetime})
}
\examples{
\dontrun{

# To be copied in the UI
mod_manage_addColumn_ui(id = "manage_addColumn_ui_1")

# To be copied in the server
mod_manage_addColumn_server(id = "manage_addColumn_ui_1",
                            infile = rective({ infile }),
                            rvs_dataset = rective({ rvs$data })
)
}

}
