% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manage_transform.R
\name{manage_transform}
\alias{manage_transform}
\alias{mod_manage_transform_ui}
\alias{mod_manage_transform_server}
\title{Column  Transformation Module}
\usage{
mod_manage_transform_ui(id)

mod_manage_transform_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to transform a column of the input dataframe
}
\examples{
\dontrun{

# To be copied in the UI
mod_manage_transform_ui(id = "manage_transform_ui_1")

# To be copied in the server
mod_manage_transform_server(id = "manage_transform_ui_1",
                            infile = reactive({ infile }),
                            rvs_dataset = reactive({ rvs$data })
)
}

}
