% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{max_scaling}
\alias{max_scaling}
\title{Normalizes with max}
\usage{
max_scaling(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
If x is a numeric variable return x/max(x)
}
\description{
Normalizes with max
}
\examples{
x <- data$Total_Power
x_max <- max_scaling(x)
head(x)
}
