% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{minmax_scaling}
\alias{minmax_scaling}
\title{Normalizes with min-max}
\usage{
minmax_scaling(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
If x is a numeric variable return (x - min(x)) / (max(x) - min(x))
}
\description{
Normalizes with min-max
}
\examples{
x <- data$Total_Power
x_minmax <- minmax_scaling(x)
head(x)
}
