% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cart.R
\name{mod_cart}
\alias{mod_cart}
\alias{mod_cart_ui_input}
\alias{mod_cart_ui_output}
\alias{mod_cart_server}
\title{CART Module}
\usage{
mod_cart_ui_input(id)

mod_cart_ui_output(id, type)

mod_cart_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{type}{a parameter describing the type of output to be displayed. The available options are:
\code{tree}, \code{cp}, \code{cm}}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to perform CART
}
\examples{
\dontrun{

# To be copied in the UI
mod_cart_ui_input(id = "cart_ui_1")
mod_cart_ui_output(id = "cart_ui_1", type = c("tree", "cp", "cm"))

# To be copied in the server
mod_cart_server(id = "cart_ui_1",
               infile = reactive({ infile }),
               rvs_dataset = reactive({ rvs$data })
)
}

}
