% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clustering.R
\name{mod_clustering}
\alias{mod_clustering}
\alias{mod_clustering_ui_input}
\alias{mod_clustering_ui_output}
\alias{mod_clustering_server}
\title{Daily Profile Clustering Module}
\usage{
mod_clustering_ui_input(id)

mod_clustering_ui_output(id)

mod_clustering_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to perform daily load profile clustering
}
\examples{
\dontrun{

# To be copied in the UI
mod_clustering_ui_input(id = "clustering_ui_1")
mod_clustering_ui_output(id = "clustering_ui_1")

# To be copied in the server
mod_clustering_server(id = "clustering_ui_1",
                      infile = reactive({ infile }),
                      rvs_dataset = reactive({ rvs$data })
)
}

}
