% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocessing_cleaning.R
\name{mod_preprocessing_cleaning}
\alias{mod_preprocessing_cleaning}
\alias{mod_preprocessing_cleaning_ui_input}
\alias{mod_preprocessing_cleaning_ui_output}
\alias{mod_preprocessing_cleaning_server}
\title{Preprocessing Module}
\usage{
mod_preprocessing_cleaning_ui_input(id)

mod_preprocessing_cleaning_ui_output(id)

mod_preprocessing_cleaning_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to preprocess datasets through different methods.
}
\examples{
\dontrun{

# To be copied in the UI
mod_preprocessing_cleaning_ui_input(id = "preprocessing_cleaning_ui_1")
mod_preprocessing_cleaning_ui_output(id = "preprocessing_cleaning_ui_1")

# To be copied in the server
mod_preprocessing_cleaning_server(id = "preprocessing_cleaning_ui_1",
                                  infile = reactive({ infile }),
                                  data_set = reactive({ rvs$data })
)
}
}
