% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocessing_outliers.R
\name{mod_preprocessing_outliers}
\alias{mod_preprocessing_outliers}
\alias{mod_preprocessing_outliers_ui_input}
\alias{mod_preprocessing_outliers_ui_output}
\alias{mod_preprocessing_outliers_server}
\title{Preprocessing Module for outliers identification}
\usage{
mod_preprocessing_outliers_ui_input(id)

mod_preprocessing_outliers_ui_output(id)

mod_preprocessing_outliers_server(id, infile = NULL, rvs_dataset)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{infile}{A reactive boolean used to understand if a dataset has been loaded on client side. It is used to disable buttons and avoids incorrect user inputs. Pass as \code{reactive({...})}.}

\item{rvs_dataset}{A reactive values dataset created from \code{reactiveValues()} and passed to the module from the external environment. Pass as \code{reactive({...})}.}
}
\description{
This module permits to identify outliers through different techniques.
The main implemented technique is the interquartile method. Through the user interface ...
}
\examples{
\dontrun{

# To be copied in the UI
mod_preprocessing_outliers_ui_input(id = "preprocessing_outliers_ui_1")
mod_preprocessing_outliers_ui_output(id = "preprocessing_outliers_ui_1")

mod_preprocessing_outliers_server(id = "preprocessing_outliers_ui_1",
                                 infile = reactive({infile}), 
                                 rvs_dataset = reactive({data_rv})  
)
}

}
