% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{zscore}
\alias{zscore}
\title{Z score normalization}
\usage{
zscore(x, na.rm = TRUE)
}
\arguments{
\item{x}{Input variable}

\item{na.rm}{If TRUE missing values are removed before calculation}
}
\value{
If x is a numeric variable return (x - mean(x)) / sd(x)
}
\description{
Normalizes with z score
}
\examples{
\dontrun{
x <- data$Total_Power
x_zscore <- zscore(x)
x
}
}
